fix an error of calculating harvest speed being divided by zero

+++ dspace-api/src/main/java/org/dspace/harvest/OAIHarvester.java
@@ -528,7 +528,7 @@ public class OAIHarvester {
         harvestRow.setHarvestStatus(HarvestedCollection.STATUS_READY);
         log.info(
             "Harvest from " + oaiSource + " successful. The process took " + timeTaken + " milliseconds. Harvested "
-                + currentRecord + " items. Harvesting speed is " + (currentRecord / (timeTaken / 1000 / 60)) +
+                + currentRecord + " items. Harvesting speed is " + (currentRecord / ((float) timeTaken / 1000 / 60)) +
                 " items/min.");
         harvestedCollectionService.update(ourContext, harvestRow);
